+++ ChangeLog.md
@@ -1,6 +1,12 @@
 # Revision history for th-abstraction
 
-## 0.1.1.0  --
+## 0.1.2.0  --
+
+* Added `arrowKCompat`
+* Added workaround for GHC 7.8 data instance eta reduction bug
+* Added kind signatures to datatypeVars
+
+## 0.1.1.0  -- 2017-05-20
 
 * Better matching of constraints generated for GADTs across GHC versions
 * Added `dataDCompat`